simplified top bar and fixed divs
