+++ playground/App.jsx
@@ -5,31 +5,15 @@ import PyatchStage from '../src/components/PyatchStage.jsx';
 import Grid from '@mui/material/Grid';
 import './App.css'
 import { PyatchCodeEditor } from '../src/components/PyatchEditor.jsx';
-import { default as PatchFileButton } from '../src/components/PatchFileButton.jsx';
-import { default as PatchFileName } from '../src/components/PatchFileName.jsx';
-import { default as PatchProjectButton } from '../src/components/PatchProjectButton.jsx';
-import { default as PatchSignOutButton } from '../src/components/PatchSignOutButton.jsx';
+import PatchTopBar from '../src/components/PatchTopBar.jsx';
 
 function App() {
 
   return (
     <PyatchProvider>
       <Grid container width={'100%'} spacing={2}>
-        <Grid container item direction = "row" xs = {8} spacing={2}>
-          <Grid item>
-            <PatchFileButton/>
-          </Grid>
-          <Grid item xs={6}>
-            <PatchFileName/>
-          </Grid>
-        </Grid>
-        <Grid container item direction = "row" xs = {4} spacing = {2} justifyContent = "flex-end">
-          <Grid item>
-            <PatchProjectButton/>
-          </Grid>
-          <Grid item>
-            <PatchSignOutButton/>
-          </Grid>
+        <Grid container item direction ="row" xs={12}>
+          <PatchTopBar/>
         </Grid>
         <Grid item xs={7}>
           <PyatchCodeEditor/>
